pr-review-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and LangGraph
langgraph==0.0.26
langchain==0.1.4
langchain-google-genai==0.0.6

# GitHub Integration
PyGithub==2.1.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
